pyproject.toml
[tool.poetry]
name = "transfile"
version = "0.1.0"
description = ""
authors = ["hsn <hsn1919810@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pywebio = "^1.8.3"
ffmpeg-python = "^0.2.0"
pywebview = {extras = ["qt"], version = "^5.1"}
flask = "^3.0.3"


[tool.poetry.group.dev.dependencies]
tqdm = "^4.66.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
